wagtail/requirements.txt
aniso8601==7.0.0
asgi-redis==1.4.3
asgiref==1.1.2
asn1crypto==1.0.0
attrs==19.2.0
autobahn==19.10.1
Automat==0.7.0
beautifulsoup4==4.6.0
certifi==2019.9.11
cffi==1.12.3
channels==1.1.8
chardet==3.0.4
colorthief==0.2.1
constantly==15.1.0
cryptography==2.7
daphne==1.4.2
Django==2.2.6
django-cors-headers==3.1.1
django-modelcluster==4.4
django-taggit==0.24.0
django-treebeard==4.3
djangorestframework==3.10.3
draftjs-exporter==2.1.7
graphene==2.1.8
graphene-django==2.6.0
graphql-core==2.2.1
graphql-relay==2.0.0
graphql-ws==0.3.0
gunicorn==19.9.0
html5lib==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
msgpack-python==0.5.6
Pillow==6.2.0
promise==2.2.1
pycparser==2.19
PyHamcrest==1.9.0
pytz==2019.2
redis==2.10.6
requests==2.22.0
Rx==1.6.1
scour==0.37
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
Unidecode==1.1.1
urllib3==1.25.6
wagtail==2.6.2
wagtail-grapple==0.3.1
wagtail-headless-preview==0.1.4
webencodings==0.5.1
Willow==1.1
zope.interface==4.6.0
